+++ contracts/Core/interface/IAssetManager.sol
@@ -4,13 +4,13 @@ pragma solidity ^0.8.0;
 interface IAssetManager {
     function executePendingDeactivations(uint32 epoch) external;
 
-    function getActiveCollections() external view returns (uint8[] memory);
+    function getActiveCollections() external view returns (uint16[] memory);
 
-    function getPendingDeactivations() external view returns (uint8[] memory);
+    function getPendingDeactivations() external view returns (uint16[] memory);
 
-    function getCollectionIndex(uint8 id) external view returns (uint8);
+    function getCollectionIndex(uint16 id) external view returns (uint16);
 
     function getNumActiveCollections() external view returns (uint256);
 
-    function getCollectionPower(uint8 id) external view returns (int8);
+    function getCollectionPower(uint16 id) external view returns (int8);
 }
